convert on delete no action to on delete cascade
When the foreign key constraint was added, I forgot to add a ON DELETE
CASCADE to the constraint. Without it, it will automatically do an ON
DELETE NO ACTION which will restrict the parent from being deleted if a
child row references it. This caused a bug where we were unable to
delete a pipeline because there were build image resource cache rows
that were owned by a job within the pipeline they were trying to delete.
By adding an ON DELETE CASCADE, deleting the pipeline will delete the
job which will now delete any build image resource caches that were for
that job.

[#6577]

Signed-off-by: Clara Fu <fclara@vmware.com>