B: support drag and drop for instanced pipelines
we needed to change how we displayed Group headers - previously, we were
just using teamName as the header text. for instance group headers, we
previously just made the `teamName` the header we wanted
(i.e. teamName ++ " / " ++ groupName). this was a pretty bad hack,
because we rely on comparing the teamName for some drag and drop logic.

to fix this, we created a new type "Section" which allows for specifying
a header that differs from the team name. this way, instance groups can
have a header like before (teamName / groupName) while still using the
correct team name for comparisons.

Signed-off-by: Aidan Oldershaw <aoldershaw@pivotal.io>
Co-authored-by: Esteban Foronda <eforonda@vmware.com>