atc: avoid task specific error in load_var
Presently when a `load_var` step is used where the `file` refers to an artifact
which is not in the registry (in my case, due to a typo), the error message
erroneously refers to a task config - for example:

```
unknown artifact source: 'xyz' in task config file path 'file.json'
```

This seems to be because we are referring to the `UnknownArtifactSourceError`
defined in `atc/task_config_source.go`, which has a task config specific
`Error()`.

Helpfully (and confusingly), there is _also_ an `UnknownArtifactSourceError`
defined in the `artifact` package with a less specific name.

As such, returning `artifact.UnknownArtifactSourceError` will return an error
message like

```
unknown artifact source: 'xyz' in file path 'file.json'
```

Signed-off-by: Kieran Gorman <kieran@kjgorman.com>