skip DNS proxy test with extra DNS server
rationale is included in the comment block - it seems to us that this
test should assert that it *can* properly lookup the address, but we
don't have too much context on this so don't want to delete the test
outright. however, it fails (i.e. successfully resolves the address)
frequently, leading to failed builds on CI

Signed-off-by: Aidan Oldershaw <aoldershaw@pivotal.io>
Co-authored-by: Esteban Foronda <eforonda@vmware.com>