re-enable k8s dns proxy test
this time with more documentation so we know what's going on and how to
debug it.

also modified the task so that it prints out slightly more useful
information. by printing out the etc/resolv.conf, we can see exactly
which dns server it's going to hit and reason about the test failure a
bit more.

Signed-off-by: Esteban Foronda <eforonda@vmware.com>
Co-authored-by: Bohan Chen <bochen@pivotal.io>