worker: Set PATH based on UID instead of container's privileged state
PATH should be set based on a users UID from the perspective of the
container, not what the UID maps to outside the container.

Matches what guardian is doing:
https://github.com/cloudfoundry/guardian/blob/bf4c737f2bc490aa82b8883e2c8964680236a8df/rundmc/processes/unix_env.go#L44-L46

Now when someone `fly intercept`'s into a container they'll have
`/sbin` on their path and be able to `apt/apk` install whatever debug
tools they need.

Signed-off-by: Taylor Silva <tsilva@pivotal.io>