generate dynamic docker-compose.yml for integration
this way, we don't need to worry about the main docker-compose.yml
drifting from the override as much, since the override contains a
complete copy of the worker service. this way, we just clone the
existing service and modify it as needed in code.

also enhances the assertion to ensure we pass a COW volume in as both
the base image and the input

Signed-off-by: Aidan Oldershaw <aoldershaw@pivotal.io>