B: fix deadlock finding resource config descriptor
we don't know exactly why, but for some reason, first SELECTing FOR
UPDATE and then UPDATEing last_referenced caused a deadlock when
concurrently setting a pipeline.  Doing things this way (UPDATE
last_referenced RETURNING ...) fixes the issue.

there's still a deadlock that occurs when concurrently updating the
pipeline and checking resources, but it's now from a different query

Signed-off-by: Aidan Oldershaw <aoldershaw@pivotal.io>
Co-authored-by: Esteban Foronda <eforonda@vmware.com>