B: fix caching when streamed volume was also streamed
this fixes a bug with #6660 - if you streamed a resource cache from
worker A -> B, and then streamed that volume from B -> C, the
worker_resource_cache on C would previously be associated with the
worker_base_resource_type on B (instead of A, which is where the volume
originated).

what this would mean is that if the base resource type on A gets
invalidated, it would not invalidate the cache on C, even though it
should. it also means that if a resource cache was streamed to a
windows/darwin worker, and we tried to stream the volume from the
windows/darwin worker to another worker, we'd get an error (worker base
resource type disappeared, or something like that)

Signed-off-by: Aidan Oldershaw <aoldershaw@pivotal.io>