B: add prototypes to database
prototypes are nearly identical to resource_types at the database layer
thus far. they will start to diverge more as we implement more of the
prototype specific behaviour (e.g. we'll want to cache their info
response).

they are also different from resource_types in that they can't currently
be used as the parent type for resources - though, this will change when
we introduce resource prototypes.

note: I didn't add a `Prototypes` method to the `CheckFactory` (the
corresponding `ResourceTypes` method is currently used by lidar to run
checks for all resource types) - this is because, with a refactor as
part of #6274, resource types won't be automatically checked by lidar.
instead, they will only be checked in builds when needed. the same will
be true for prototypes, so I won't add periodic checking of prototypes
yet.

Signed-off-by: Aidan Oldershaw <aoldershaw@pivotal.io>