atc: across step logs errors
Build errors (different from failures in regards to the state of a step
within a build) are now surfaced at the across step level as well as
within the sub-step(s) where the error originates. This also ensures
that web UI correctly highlights the state of the across step.

We wanted to expand the test suite to test the get, put, and task step
since we noticed this wasn't tested anywhere. I struggled to figure out
how to make this happen though as I kept running into issues with
getting this steps to intentionally error. The scheduler is too smart
now!

Signed-off-by: Taylor Silva <tsilva@pivotal.io>
Co-authored-by: Bohan Chen <bochen@pivotal.io>