inline upstream/downstream logic
in order to support the edge case where a build outputs the same
resource version as one of its inputs, the child of a build node cannot
be the same as the parent. in order to make this work, certain
guarantees around how the tree is constructed needs to be made.

if we can guarantee that the tree is constructed from the leaf up, we
are able to simply remove the resource version from its children's
children. however, this still leaves the problem of how to look at
grandchildren

Signed-off-by: Bohan Chen <bochen@pivotal.io>
Co-authored-by: Bohan Chen <bochen@pivotal.io>