use fixed name for upgrade/downgrade test
the upgrade/downgrade tests could previously stall out because the
worker name changed, but the ATC hadn't realized that the old worker was
stalled before it tried to schedule a task on it, and then gets stuck
retrying communication with either garden or baggageclaim (not sure
which) on the missing worker.

this is a bug with the suite that currently exists on master - e.g. see
https://ci.concourse-ci.org/teams/main/pipelines/concourse/jobs/integration/builds/247#L60c1769f:451

Signed-off-by: Aidan Oldershaw <aoldershaw@pivotal.io>