B: avoid hitting join_collapse_limit
adding prototypes to the "scan build" query pushed the number of joins
over the join_collapse_limit (defaults to 8) for certain queries,
causing postgres to generate suboptimal execution plans (taking
~5minutes for a simple query, since it ended up doing seq scans for all
tables).

this commit removes the joins for resource_types and prototypes. these
joins were just to get the names of the resource_type/prototype for use
in the syslog tag, tracing, and logs. we no longer add this information
to these sinks. however, with @clarafu's refactor in #7176, we won't
generate builds for resource_types or prototypes automatically (i.e. in
lidar), as they will be checked as part of the checks for resources (or
steps in a job build), so hopefully this isn't too big a deal.

in the long run, a better solution may be to store general metadata on
the builds table so we don't have to do so many joins in the first place
- but for now, let's just get rid of these less interesting ones to
exceeding the limit

Signed-off-by: Aidan Oldershaw <aoldershaw@pivotal.io>
Co-authored-by: Alex Suraci <suraci.alex@gmail.com>