B: remove prototype_id from builds
with concourse/concourse#7176, I realize we don't gain much from keeping
track of the prototype_id or resource_type_id on the builds table, as
the only time we will ever create checks for them is when we manually
fly check-{proto,resource-}type. this ID is used:

1. to ensure we don't create unnecessary checks (for non-manually
   triggered checks)
2. to GC old checks (i.e. where there is already a newer finished
   check).
3. to emit build events to the check_build_events partition for
   optimized GC

however, given that checks for these types will be limited to only
manually triggered checks, this is no longer really worthwhile.

1. all checks are manually triggered anyway, so no difference
2. there will be far fewer checks so GC'ing likely won't be worthwhile.
   if we do want to GC these checks, however, we could add a time-based
   GC (e.g. remove check builds older than 5 days)
3. same as 2

Signed-off-by: Aidan Oldershaw <aoldershaw@pivotal.io>