B: allow interpolation in the across step
this commit enables more dynamic iteration with the across step by
interpolating the var values at runtime. the plans of the substeps are
then emitted as a build event, along with their corresponding
combination of var values, where it is then rendered by the UI.

note that for consistency/simplicity, even if the list of var values
could be known at plan time (since they are all static), we only
construct and emit the substeps at runtime. this means the across step
can't be expanded (to display the substep headers) until it starts
executing.

it also fixes a bug with the across step where if a substep was
highlighted, we wouldn't properly expand the across step header on page
load.

Signed-off-by: Aidan Oldershaw <aoldershaw@pivotal.io>