B: support queueing notification bus
...and support receiving string payloads from the bus

The current behaviour of the notification bus when a new notification
comes in is:

Try to send that notification to each Go chan:
* If that chan is empty, send along the chan
* If that chan already has a notification waiting to be received, do
  nothing

This is nice because all we previously cared about is that the consumer
is told that something changed. However, if we want to start receiving
data on the notification bus (e.g. triggers that include the ID of the
element that was changed), we'll need to queue several events in case
the consumer isn't fast enough to drain them.

Note that I tagged this commit as behaviour (since it introduces new
behaviour that can be used), but it doesn't yet change anything since
nothing is making use of it - everything is still using the "don't
queue notifications" behaviour, and nothing sends any payloads.

Signed-off-by: Aidan Oldershaw <aoldershaw@pivotal.io>