B: cache workers in memory
the motivation for this change is to optimize scheduling, since the
WorkerFactory methods (esp. Workers and BuildContainersCountPerWorker)
get called every time a container is scheduled. on hush-house, these two
queries consume a ton of CPU time.

this commit adds a new WorkerCache component that tracks the current
workers (and their corresponding container counts) in memory.
WorkerFactory shells out to WorkerCache when it lists or looks up
workers [1], and when it lists the build containers count.

WorkerCache keeps its in-memory data relatively fresh by listening to
changes in these tables via postgres TRIGGERs. when a worker is
INSERTed, UPDATEd, or DELETEd, the WorkerCache will update its internal
state accordingly. for UPDATE/INSERT, this requires making a DB query
for the single worker that changed.

similarly, we track INSERTs/DELETEs for containers. here, however, we
don't need to do another query - we just increment/decrement a counter.

it's possible that the data can still get out of sync, since it's
possible to miss updates on the notification bus. for this, we
periodically refresh the state once every minute so we don't stay out of
sync for too long.

[1] the exception here is FindWorkersForContainerByOwner, which still
makes a query. we *could* store the list of containers in memory as
well, but it doesn't seem to be a bottleneck in the DB so it didn't seem
worth it.

Signed-off-by: Aidan Oldershaw <aoldershaw@pivotal.io>