fix testflight failures
Some of the tests were testing that the resource type version does not
exist is errored but we no longer throw that error anymore because we
run a check everytime for any parent resource types. The other test that
was failing was testing that the resource/resource type gets printed
when it is checked recursively but that also does not exist anymore
because we are no longer checking recursively through fly (which is the
thing that printed that message).

[#6274]

Signed-off-by: Clara Fu <fclara@vmware.com>